requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
huey==2.3.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
